requirements.txt
appdirs>=1.4.3
Appium-Python-Client==0.34
argh>=0.26.2
astroid>=2.0.1
autopep8>=1.4
baron>=0.6.6
behave>=1.2.6
certifi>=2018.4.16
chardet>=3.0.4
click>=6.7
colorama>=0.3.9
cssselect>=1.0.3
decorator>=4.3.0
docopt>=0.6.2
Flask>=1.0.2
Flask-Login>=0.4.1
getgauge>=0.3.3
golem-framework>=0.5.0
grpcio>=1.14.0
helper>=2.4.2
howdoi>=1.1.13
idna>=2.7
invoke>=1.1.0
isort>=4.3.4
itsdangerous>=0.24
Jinja2>=2.10
jsonpath-rw>=1.4.0
lazy-object-proxy>=1.3.1
lxml>=4.2.3
MarkupSafe>=1.0
mccabe>=0.6.1
nose>=1.3.7
packaging>=17.1
parse>=1.8.4
parse-type>=0.4.2
path.py>=11.0.1
pathtools>=0.1.2
Paver>=1.3.4
peru>=1.1.4
pluggy>=0.6.0
ply>=3.11
protobuf>=3.6.0
ptvsd>=3.0.0
py>=1.5.4
py-webdriver-manager>=0.0.1
pycodestyle>=2.4.0
Pygments>=2.2.0
PyHamcrest>=1.9.0
pylint>=2.0.1
pyparsing>=2.2.0
pyquery>=1.4.0
PyYAML>=3.13
redbaron>=0.6.3
requests>=2.19.1
requests-cache>=0.4.13
robot>=20071211
robotframework>=3.0.4
robotframework-selenium2library>=3.0.0
robotframework-seleniumlibrary>=3.2.0
rply>=0.7.6
selenium>=3.13.0
six>=1.11.0
tox>=3.1.2
tqdm>=4.23.1
typed-ast>=1.1.0
urllib3>=1.23
virtualenv>=16.0.0
watchdog>=0.8.3
Werkzeug>=0.14.1
wrapt>=1.10.11
schedule>=0.6.0
